Refactor : 크롤링 리펙토링
