+++ src/main/java/com/example/newneekclone/domain/news/entity/News.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 @RequiredArgsConstructor
-public class News extends Timestamped{
+public class News {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -30,8 +30,8 @@ public class News extends Timestamped{
     @Column(name = "tags", columnDefinition = "TEXT")
     String tags;
 
-    @Column(name = "create_date")
-    LocalDateTime create_date;
+    @Column(name = "date")
+    LocalDateTime date;
 
     @Column(name = "url")
     String url;
@@ -42,26 +42,14 @@ public class News extends Timestamped{
     @Column(name = "videoURL")
     String videoUrl;
 
-//    @OneToOne(mappedBy = "news", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-//    private Images images;
-
     public News(String allTag, String title, String category, String link, String content, LocalDateTime date, String imageUrl, String videoUrl) {
         this.tags = allTag;
         this.title = title;
-        this.category= category;
+        this.category = category;
         this.url = link;
         this.content = content;
-        this.create_date = date;
+        this.date = date;
         this.imageUrl = imageUrl;
         this.videoUrl = videoUrl;
     }
-
-//    public News(String title, String category, String allContents, String date, String allTag, String text) {
-//        this.title = title;
-//        this.category= category;
-//        this.content = allContents;
-//        this.date = date;
-//        this.tags = allTag;
-//
-//    }
 }
