--- src/main/java/com/example/newneekclone/domain/news/entity/Timestamped.java
@@ -1,17 +1,0 @@
-package com.example.newneekclone.domain.news.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AutoCloseable.class)
-public abstract class Timestamped {
-    @CreatedDate
-    @Column(updatable = false)
-    @Temporal(TemporalType.TIMESTAMP)
-    private LocalDateTime createdDate;
-}
