--- src/main/java/com/example/newneekclone/domain/news/repository/ImagesRepository.java
@@ -1,9 +1,0 @@
-//package com.example.newneekclone.domain.news.repository;
-//
-//import com.example.newneekclone.domain.news.entity.Images;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.stereotype.Repository;
-//
-//@Repository
-//public interface ImagesRepository extends JpaRepository<Images, Long> {
-//}
